Crate funding_trading_bridge_smart_contract

Funding Trading Bridge Smart Contract

This contract uses Cosmwasm’s provided architecture in conjunction with Provwasm to create a wasm smart contract that can be deployed to and interact with the Provenance Blockchain.

This contract is designed to facilitate movement of funds from two different restricted Markers. It allows one-to-one trades for the two markers, accounting for differences in artificially described precisions for the coin counts in the marker denoms.

Modules

contract

The entrypoint for all external commands sent to the compiled wasm.

execute

All code and functions pertaining to the execute entrypoint. Contains all execution routes used by the contract file.

instantiate

All code and functions pertaining to the instantiate entrypoint. Contains the functionality used in the [contract file](crate::contract] to instantiate an instance of the contract.

migrate

All code and functions pertaining to the migrate entrypoint. Contains the functionality used in the contract file to perform a migration to a new version.

query

All code and functions pertaining to the query entrypoint. Contains the functionality used in the contract file to perform a query.

store

All code and functions pertaining to interacting with mutable contract data storage. Contains all type definitions and functionality for interacting with contract internal storage.

types

All globally-defined structs used by functions throughout the contract. Contains all types and base functionality used to construct the logic of the contract.

util

Utility functions and traits adopted by various aspects of the contract. Additional functionality that does not strictly belong to a category.