Module funding_trading_bridge_smart_contract::execute::withdraw_trading

This execution route converts the trading marker denom to the deposit marker denom by transferring the trading marker denom from the sender to the trading marker itself, burning the received values, and then returning deposit marker denom to the sender’s account.

Functions

withdraw_trading

Invoked via the contract’s execute functionality. The function will attempt to pull trade_amount of the trading marker’s denom from the sender’s account with a marker transfer, discern how much of the deposit denom to which the submitted amount is equivalent, transfer that amount to the sender, and then burn the exchanged trading marker denom.