Function funding_trading_bridge_smart_contract::store::contract_state::get_contract_state_v1

pub fn get_contract_state_v1(
    storage: &dyn Storage
) -> Result<ContractStateV1, ContractError>

Fetches the current contract instance of contract state. This call should never fail because the state is set on contract instantiation, but an error will be returned if store communication fails.

Parameters

storage An immutable instance of the contract storage value, allowing internal store data fetches.