Enum funding_trading_bridge_smart_contract::types::msg::ExecuteMsg

pub enum ExecuteMsg {
    AdminUpdateAdmin {
        new_admin_address: String,
    },
    AdminUpdateDepositRequiredAttributes {
        attributes: Vec<String>,
    },
    AdminUpdateWithdrawRequiredAttributes {
        attributes: Vec<String>,
    },
    FundTrading {
        trade_amount: Uint128,
    },
    WithdrawTrading {
        trade_amount: Uint128,
    },
}

All defined paylods to be used when executing routes on this contract instance.

Variants

AdminUpdateAdmin

Fields

new_admin_address: String

A bech32 address to use as the new administrator of the contract.

A route that swaps the current value in the contract state for the admin to the provided value.

AdminUpdateDepositRequiredAttributes

Fields

attributes: Vec<String>

The new attributes that will be set in the contract state’s required_deposit_attributes property upon successful execution.

A route that sets a new collection of attribute names required when an account deposits their deposit denom into the contract via the fund_trading execution route.

AdminUpdateWithdrawRequiredAttributes

Fields

attributes: Vec<String>

The new attributes that will be set in the contract state’s required_withdraw_attributes property upon successful execution.

A route that sets a new collection of attribute names required when an account withdraws their deposit denom from the contract via the withdraw_trading execution route.

FundTrading

Fields

trade_amount: Uint128

The amount of the deposit marker to pull from the sender’s account in exchange for trading denom.

A route that will attempt to pull the trade amount of the deposit marker’s denom from the sender’s account with a marker transfer, discern how much of the trading denom to which the submitted amount is equivalent, and then mint and withdraw the equivalent amount into the sender’s account.

WithdrawTrading

Fields

trade_amount: Uint128

The amount of the trading marker to pull from the sender’s account in exchange for deposit denom.

A route that will attempt to pull the trade amount of the trading marker’s denom from the sender’s account with a marker transfer, discern how much of the deposit denom to which the submitted amount is equivalent, transfer that amount to the sender, and then burn the exchanged trading marker denom.