Module funding_trading_bridge_smart_contract::util::provenance_utils

Utility functions for interacting with Provenance Blockchain resources.

Functions

check_account_has_all_attributes

Ensures that the target account has all the specified attributes. Does not check for valid attribute body contents.

check_account_has_enough_denom

Ensures that the target account holds enough of the target denom name by verifying their balances in the bank module.

get_marker_address_for_denom

Fetches the bech32 address associated with the marker account for the given denomination.

msg_bind_name

Generates a name bind msg that will properly assign the given name value to a target address. Assumes the parent name is unrestricted or that the contract has access to bind a name to the parent name.