Function funding_trading_bridge_smart_contract::util::provenance_utils::msg_bind_name

pub fn msg_bind_name<S1: Into<String>, S2: Into<String>>(
    name: S1,
    bind_to_address: S2,
    restricted: bool
) -> Result<MsgBindNameRequest, ContractError>

Generates a name bind msg that will properly assign the given name value to a target address. Assumes the parent name is unrestricted or that the contract has access to bind a name to the parent name.

Parameters

name The dot-qualified name to use on-chain for name binding. Ex: myname.sc.pb will generate a msg that binds “myname” to the existing parent name “sc.pb”.

bind_to_address The bech32 address to which the name will be bound.

restricted If true, the name will be bound as a restricted name, preventing future name bindings from using it as a parent name.